+++ types/view/three.ts
@@ -242,8 +242,15 @@ interface Quotations extends View {
   strip: () => View
 }
 
-// only overloads .json()
-interface Adjectives extends View { }
+interface Adjectives extends View {
+  /** get the words describing this adjective */
+  adverbs: () => View
+  /** 'quick' -> 'quickly' */
+  toAdverb: (n?: number) => View
+  /** 'quick' -> 'quickness' */
+  toNoun: (n?: number) => View
+}
+
 // only overloads .json()
 interface Adverbs extends View { }