+++ src/2-two/preTagger/model/regex/regex-text.js
@@ -11,4 +11,7 @@ export default [
   // ending-apostrophes
   [/.{3}[lkmnp]in['‘’‛‵′`´]$/, 'Gerund', "chillin'"],
   [/.{4}s['‘’‛‵′`´]$/, 'Possessive', "flanders'"],
+
+  //from https://www.regextester.com/106421
+  [/^([\u00a9\u00ae\u2319-\u3300]|\ud83c[\ud000-\udfff]|\ud83d[\ud000-\udfff]|\ud83e[\ud000-\udfff])/, 'Emoji', 'emoji-range']
 ]