+++ changelog.md
@@ -12,9 +12,10 @@ While all _Major_ releases should be reviewed, our only two _large_ releases are
 
 -->
 
-#### 14.0.1 [March 2021]
+#### 14.1.0 [March 2021]
 - **[fix]** - client-side export format for plugins
 - **[new]** - more adjective transformation methods
+- **[new]** - emoji + emoticon tagger
 
 ### 🚨 14.0.0 🚨 &nbsp; [March 2022]
 Major release - see [Release Notes](https://github.com/spencermountain/compromise/releases/tag/14.0.0) for full details